Replace commander with yargs library
