+++ package.json
@@ -1,16 +1,16 @@
 {
   "name": "artifactoid",
-  "version": "0.1.6",
+  "version": "0.2.0",
   "description": "A command line tool to retrieve the URI of the latest artifact from an Artifactory repository",
-  "main": "index.js",
+  "preferGlobal": true,
+  "bin": {
+    "artifactoid": "src/cli.js"
+  },
   "scripts": {
     "lint": "eslint ./src/index.js",
     "coverage": "jest --coverage",
     "coveralls": "cat ./coverage/lcov.info | coveralls",
     "test": "yarn lint && yarn coverage && yarn coveralls"
-  },
-  "bin": {
-    "artifactoid": "./src/index.js"
   },
   "repository": {
     "type": "git",
@@ -25,9 +25,10 @@
   "dependencies": {
     "base-64": "^0.1.0",
     "chalk": "^1.1.3",
-    "commander": "^2.9.0",
     "natural-sort": "^1.0.0",
-    "node-fetch": "^1.6.3"
+    "node-fetch": "^1.6.3",
+    "open": "^0.0.5",
+    "yargs": "^8.0.1"
   },
   "devDependencies": {
     "babel-jest": "^20.0.0",
